Page List: Move modal to it's own file
